Move view refresh logic into ClickhouseTableMetrics
_collect_view_refresh_metrics and _handle_view_refreshes_error now live
on ClickhouseTableMetrics and run inside run_job(), sharing the same
collection interval and DBMAsyncJob lifecycle as table size gauges.
State flags and query/status-map constants move to table_metrics.py.
clickhouse.py loses the duplicate if-table_metrics block, the two
module-level constants, the class attribute, and the three init flags.
